Add access control

+++ lib/pintelier_web/live/group_live/index.ex
@@ -1,5 +1,5 @@
 defmodule PintelierWeb.GroupLive.Index do
-alias Pintelier.Groups.GroupMember
+  alias Pintelier.Groups.GroupMember
   use PintelierWeb, :live_view
 
   alias Pintelier.Groups
@@ -16,19 +16,31 @@ alias Pintelier.Groups.GroupMember
   end
 
   defp apply_action(socket, :edit, %{"id" => id}) do
-    socket
-    |> assign(:page_title, "Edit Group")
-    |> assign(:group, Groups.get_group!(id))
+    group = Groups.get_group!(id)
+
+    if Groups.can_edit_group?(socket.assigns.current_user, group) do
+      socket
+      |> assign(:page_title, "Edit Group")
+      |> assign(:group, Groups.get_group!(id))
+    else
+      socket
+      |> put_flash(:error, "Unauthorized")
+      # Redirect because it shouldn't be possible to get here without client manipulation
+      |> redirect(to: ~p"/groups")
+    end
   end
 
   defp apply_action(socket, :new, _params) do
     socket
     |> assign(:page_title, "New Group")
-    |> assign(:group, %Group{members: [
-      socket.assigns.current_user
-    ], group_members: [
-      %GroupMember{user: socket.assigns.current_user, authorization: :admin}
-    ]})
+    |> assign(:group, %Group{
+      members: [
+        socket.assigns.current_user
+      ],
+      group_members: [
+        %GroupMember{user: socket.assigns.current_user, authorization: :admin}
+      ]
+    })
   end
 
   defp apply_action(socket, :index, _params) do
@@ -45,8 +57,13 @@ alias Pintelier.Groups.GroupMember
   @impl true
   def handle_event("delete", %{"id" => id}, socket) do
     group = Groups.get_group!(id)
-    {:ok, _} = Groups.delete_group(group)
-
-    {:noreply, stream_delete(socket, :groups, group)}
+    if Groups.can_delete_group?(socket.assigns.current_user, group) do
+      {:ok, _} = Groups.delete_group(group)
+      {:noreply, stream_delete(socket, :groups, group)}
+    else
+      socket
+      |> put_flash(:error, "Unauthorized")
+      |> redirect(to: "~p/groups")
+    end
   end
 end
